+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 # CHANGELOG
 
 <p>
+<b>Version 2.1.27</b> - July 2023<br/>
+- Trashed some old unuseful code and status options.<br/>
+- Keep moving help to the help box of node-red.<br/>
+</p>
+<p>
 <b>Version 2.1.26</b> - July 2023<br/>
 - Hue Light:  fixed brightness states issue.<br/>
 </p>